http: u32 -> usize for [very] large content length

+++ src/network/HttpClient.zig
@@ -2722,9 +2722,9 @@ pub const Transfer = struct {
         return .{ .curl = .{ .conn = self._conn.? } };
     }
 
-    pub fn getContentLength(self: *const Transfer) ?u32 {
+    pub fn getContentLength(self: *const Transfer) ?usize {
         const cl = self.getContentLengthRawValue() orelse return null;
-        return std.fmt.parseInt(u32, cl, 10) catch null;
+        return std.fmt.parseInt(usize, cl, 10) catch null;
     }
 
     fn getContentLengthRawValue(self: *const Transfer) ?[]const u8 {
